+++ server/controllers/PurchaseRequisitionsController.ts
@@ -1,10 +1,15 @@
 import * as express from "express";
 import PurchaseRequisitionService from "../services/PurchaseRequisitionService";
 import IPurchaseRequisitionService from "../services/interfaces/PurchaseRequisitionService";
-import { httpStatus } from "../utils/Utils";
+import { httpStatus } from "../common/Utils";
+import AirtableService from "../services/AirtableService";
+import MapHelper from "../common/MapHelper";
 
 const purchaseRequisitionsController = express.Router();
-const purchaseRequisitionsService: IPurchaseRequisitionService = new PurchaseRequisitionService();
+const purchaseRequisitionsService: IPurchaseRequisitionService = new PurchaseRequisitionService(
+    new AirtableService(),
+    new MapHelper()
+);
 
 purchaseRequisitionsController.route("/getPurchaseRequisitions").post(
     async (req, res): Promise<void> => {
